# docs/office_openapi.yaml
openapi: 3.0.3
info:
  title: Office API
  version: 1.0.0
  description: API para consulta de categorias e tipos de categorias do Office
  contact:
    name: Bruno Jet
    email: brunojet@example.com
servers:
  - url: http://localhost:8080
paths:
  /categorias:
    get:
      tags:
        - Categoria
      summary: Lista categorias
      description: Retorna uma lista paginada de categorias
      operationId: listCategorias
      parameters:
        - name: page
          in: query
          description: Página desejada
          required: false
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          description: Tamanho da página
          required: false
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Lista de categorias
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Categoria'
        '500':
          description: Erro interno
  /tipo_categorias:
    get:
      tags:
        - TipoCategoria
      summary: Lista tipos de categoria
      description: Retorna uma lista paginada de tipos de categoria
      operationId: listTipoCategorias
      parameters:
        - name: page
          in: query
          description: Página desejada
          required: false
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          description: Tamanho da página
          required: false
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Lista de tipos de categoria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TipoCategoria'
        '500':
          description: Erro interno
components:
  schemas:
    Categoria:
      type: object
      properties:
        id:
          type: integer
          format: int64
        id_tipo_categoria:
          type: integer
          format: int64
        tipo_categoria:
          $ref: '#/components/schemas/TipoCategoria'
        id_pai:
          type: integer
          format: int64
          nullable: true
          pai:
            $ref: '#/components/schemas/CategoriaPai'
            nullable: true
    TipoCategoria:
      type: object
      properties:
        id:
          type: integer
          format: int64

    CategoriaPai:
      type: object
      description: Categoria pai simplificada
      properties:
        id:
          type: integer
          format: int64
        id_tipo_categoria:
          type: integer
          format: int64